#ELLIE-217 마법사 스켈레톤 공격시 플레이어 시야 고정 수정
