+++ Assets/Prefabs/Monster/Stage1/AdventureSkeleton/AdventureSkeleton.prefab
@@ -1485,7 +1485,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &1068597257678563727
 Transform:
   m_ObjectHideFlags: 0
