+++ Assets/Prefabs/Monster/Stage1/AdventureSkeleton/AdventureSkeletonBehaviourTree.asset
@@ -139,8 +139,8 @@ MonoBehaviour:
     - rid: 6937830456686542857
     - rid: 6937830456686542858
     - rid: 6937830456686542871
-  viewPosition: {x: 1133.5, y: -9.5, z: 0}
-  viewScale: {x: 0.3759371, y: 0.3759371, z: 1}
+  viewPosition: {x: 763, y: 88.5, z: 0}
+  viewScale: {x: 0.25, y: 0.25, z: 1}
   references:
     version: 2
     RefIds:
