+++ Assets/Prefabs/Monster/Stage1/CaveBat/CaveBat.prefab
@@ -908,7 +908,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &6863236723056871730
 Transform:
   m_ObjectHideFlags: 0
