+++ Assets/Prefabs/Monster/Stage1/CaveBat/CaveBatBehaviourTree.asset
@@ -119,7 +119,7 @@ MonoBehaviour:
     - rid: 6937830615371481108
     - rid: 6937830615371481109
     - rid: 6937830615371481110
-  viewPosition: {x: 5285, y: -437.5, z: 0}
+  viewPosition: {x: 1820, y: -564.5, z: 0}
   viewScale: {x: 0.86956525, y: 0.86956525, z: 1}
   references:
     version: 2
@@ -670,7 +670,7 @@ MonoBehaviour:
         drawGizmos: 0
         playerPos:
           reference:
-            rid: -2
+            rid: 5827610086317228070
           defaultValue: {x: 0, y: 0, z: 0}
     - rid: 5827610152718303232
       type: {class: Sequencer, ns: TheKiwiCoder, asm: Assembly.Ellie}
@@ -843,7 +843,7 @@ MonoBehaviour:
         drawGizmos: 0
         playerPos:
           reference:
-            rid: -2
+            rid: 5827610086317228070
           defaultValue: {x: 0, y: 0, z: 0}
     - rid: 6937830456686542868
       type: {class: MonsterLookPlayer, ns: , asm: Assembly.Ellie}
@@ -858,7 +858,7 @@ MonoBehaviour:
         drawGizmos: 0
         playerPos:
           reference:
-            rid: -2
+            rid: 5827610086317228070
           defaultValue: {x: 0, y: 0, z: 0}
     - rid: 6937830456686542869
       type: {class: BooleanKey, ns: TheKiwiCoder, asm: Assembly.Ellie}
