+++ Assets/Prefabs/Monster/Stage1/GuildguardSkeleton/GuildguardSkeleton.prefab
@@ -202,7 +202,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &3915295574766006576
 Transform:
   m_ObjectHideFlags: 0
@@ -309,7 +309,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 58c6f9b140a52a14881ec09eaebb66eb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  behaviourTree: {fileID: 11400000, guid: 77862248fcc864cab8d9a13ceb088336, type: 2}
+  behaviourTree: {fileID: 11400000, guid: be73565d44d1f4e7cab99e8b770245b6, type: 2}
   validate: 1
   blackboardOverrides: []
   references:
@@ -444,7 +444,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 10c1b8bd8441641d681ab1c2ad921155, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  data: {fileID: 11400000, guid: b5eb79329df1549d3a18c06cb09e3c91, type: 2}
+  data: {fileID: 11400000, guid: 6c591b4d7375d4579807665bb2bcc69f, type: 2}
 --- !u!114 &1126190619
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -457,7 +457,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: f84f7f233335045a68404cfb038289c3, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  data: {fileID: 11400000, guid: c820c946f59004cb3832834d4cb87342, type: 2}
+  data: {fileID: 11400000, guid: b19edce2b991d45c7a6cfaec097ec8b9, type: 2}
 --- !u!1 &557302774346244468
 GameObject:
   m_ObjectHideFlags: 0
