+++ Assets/Prefabs/Monster/Stage1/NormalSkeleton/NormalSkeleton.prefab
@@ -3218,7 +3218,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 10c1b8bd8441641d681ab1c2ad921155, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  data: {fileID: 0}
+  data: {fileID: 11400000, guid: 299ce5f5ac0564aa9bd73d47c2d4a2e4, type: 2}
 --- !u!1 &9178374280581418941
 GameObject:
   m_ObjectHideFlags: 0
