+++ Assets/Prefabs/Monster/Stage1/NormalSkeleton/NormalSkeletonParticleData.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 299ce5f5ac0564aa9bd73d47c2d4a2e4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
