+++ Assets/Prefabs/Monster/Stage1/NormalSkeleton/NormalSkeletonSkeletonBehaviourTree.asset
@@ -102,6 +102,8 @@ MonoBehaviour:
   - rid: 6937830886797737989
   - rid: 6937830886797737990
   - rid: 6937830886797737991
+  - rid: 6937830886797737993
+  - rid: 6937830886797737994
   treeState: 0
   blackboard:
     keys:
@@ -129,8 +131,8 @@ MonoBehaviour:
     - rid: 6937830456686542857
     - rid: 6937830456686542858
     - rid: 6937830456686542871
-  viewPosition: {x: -239, y: -249.5, z: 0}
-  viewScale: {x: 0.7561437, y: 0.7561437, z: 1}
+  viewPosition: {x: 2959, y: -448, z: 0}
+  viewScale: {x: 0.86956525, y: 0.86956525, z: 1}
   references:
     version: 2
     RefIds:
@@ -231,7 +233,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: fdc5cff4db3a84fa48f982b76a45a3c7
-        position: {x: -2700, y: 585}
+        position: {x: -3015, y: 585}
         blackboard:
           keys: []
         description: Attacks
@@ -245,7 +247,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: e52811ae9daed4e1eb83d1290970ec3d
-        position: {x: -3270, y: 720}
+        position: {x: -3585, y: 720}
         blackboard:
           keys: []
         description: RunToPlayer
@@ -262,7 +264,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 66312dc16c5ab4f63ac0cbda231190be
-        position: {x: -3675, y: 855}
+        position: {x: -3975, y: 855}
         blackboard:
           keys: []
         description: 
@@ -289,7 +291,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 5eb7fdc77efdb4848a31a0da9f39c3ac
-        position: {x: -3120, y: 855}
+        position: {x: -3435, y: 855}
         blackboard:
           keys: []
         description: 
@@ -304,7 +306,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: d4fbf4534dcd34430ad214f2879e66cf
-        position: {x: -2250, y: 720}
+        position: {x: -2415, y: 720}
         blackboard:
           keys: []
         description: MeleeAttack
@@ -312,6 +314,8 @@ MonoBehaviour:
         children:
         - rid: 5827610086317228044
         - rid: 5827610129303863303
+        - rid: 6937830886797737993
+        - rid: 6937830886797737994
         - rid: 6937830705275600954
         - rid: 5827610181934514197
         - rid: 5827610086317228046
@@ -322,7 +326,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: e97593efb4bd34a9288b132a303fc66d
-        position: {x: -2760, y: 855}
+        position: {x: -3075, y: 855}
         blackboard:
           keys: []
         description: 
@@ -349,7 +353,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 5cc5555f37fb94b67baf5caa4c84622d
-        position: {x: -2070, y: 855}
+        position: {x: -2085, y: 855}
         blackboard:
           keys: []
         description: 
@@ -846,7 +850,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 9e8421b9478ee498db3e0eb23307d68e
-        position: {x: -2580, y: 855}
+        position: {x: -2895, y: 855}
         blackboard:
           keys: []
         description: 
@@ -857,7 +861,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: bb46c627243974245a81ab56605ea555
-        position: {x: -1920, y: 855}
+        position: {x: -1935, y: 855}
         blackboard:
           keys: []
         description: 
@@ -910,7 +914,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 3ff18f65cd91145a6923f6adc2b049aa
-        position: {x: -3480, y: 855}
+        position: {x: -3795, y: 855}
         blackboard:
           keys: []
         description: 
@@ -929,7 +933,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: a9b72822e58e5455ea148fb29c4f1348
-        position: {x: -2970, y: 855}
+        position: {x: -3285, y: 855}
         blackboard:
           keys: []
         description: 
@@ -948,7 +952,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: afb2edf6513894559b8482e96ac10960
-        position: {x: -2280, y: 855}
+        position: {x: -2295, y: 855}
         blackboard:
           keys: []
         description: 
@@ -1372,7 +1376,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: a4a35deb9a5bd4340a09046a6d9eeec2
-        position: {x: -3270, y: 855}
+        position: {x: -3585, y: 855}
         blackboard:
           keys: []
         description: 
@@ -1395,7 +1399,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 850456f5b20214fbebb5cbc871a95034
-        position: {x: -2430, y: 855}
+        position: {x: -2445, y: 855}
         blackboard:
           keys: []
         description: 
@@ -1815,7 +1819,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 4330c184fae434e53960f8f883fcfe66
-        position: {x: -2700, y: 435}
+        position: {x: -3015, y: 435}
         blackboard:
           keys: []
         description: 
@@ -1829,7 +1833,7 @@ MonoBehaviour:
         state: 0
         started: 0
         guid: 2c24c4137d33a47a7a389bafd0974fda
-        position: {x: -2955, y: 570}
+        position: {x: -3270, y: 570}
         blackboard:
           keys: []
         description: 
@@ -1853,3 +1857,34 @@ MonoBehaviour:
           reference:
             rid: 5827610086317228070
           defaultValue: {x: 0, y: 0, z: 0}
+    - rid: 6937830886797737993
+      type: {class: PlayParticle, ns: , asm: Assembly.Ellie}
+      data:
+        state: 0
+        started: 0
+        guid: 855fd187c783e43cdbbe9d7a31838b0c
+        position: {x: -2745, y: 855}
+        blackboard:
+          keys: []
+        description: 
+        drawGizmos: 0
+        particleType:
+          reference:
+            rid: -2
+          defaultValue: 0
+    - rid: 6937830886797737994
+      type: {class: Wait, ns: TheKiwiCoder, asm: Assembly.Ellie}
+      data:
+        state: 0
+        started: 0
+        guid: 64e2c2d40c5a241b48c0178e683821cb
+        position: {x: -2595, y: 855}
+        blackboard:
+          keys: []
+        description: 
+        drawGizmos: 0
+        timeValue:
+          reference:
+            rid: -2
+          defaultValue: 0.4
+        duration: 1
