+++ Assets/Prefabs/Monster/Stage1/WizardSkeleton/WizardSkeleton.prefab
@@ -1736,7 +1736,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &2104673137817031364
 Transform:
   m_ObjectHideFlags: 0
