+++ Assets/Scenes/Monsters/Stage1 1.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.18028378, g: 0.22571412, b: 0.30692285, a: 1}
+  m_IndirectSpecularColor: {r: 0.18028326, g: 0.22571333, b: 0.30692202, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -17109,15 +17109,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 6863236723056871730, guid: 9bfcdf65613c4478d833460aa194c908, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -6.0335236
+      value: 131.20805
       objectReference: {fileID: 0}
     - target: {fileID: 6863236723056871730, guid: 9bfcdf65613c4478d833460aa194c908, type: 3}
       propertyPath: m_LocalPosition.y
-      value: -18.188068
+      value: 1.0030727
       objectReference: {fileID: 0}
     - target: {fileID: 6863236723056871730, guid: 9bfcdf65613c4478d833460aa194c908, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -8.605867
+      value: 25.985857
       objectReference: {fileID: 0}
     - target: {fileID: 6863236723056871730, guid: 9bfcdf65613c4478d833460aa194c908, type: 3}
       propertyPath: m_LocalRotation.w
@@ -17959,15 +17959,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 1068597257678563727, guid: ef1dd0845aa6a4cef9e999043447d039, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -14.676422
+      value: 122.565155
       objectReference: {fileID: 0}
     - target: {fileID: 1068597257678563727, guid: ef1dd0845aa6a4cef9e999043447d039, type: 3}
       propertyPath: m_LocalPosition.y
-      value: -18.194168
+      value: 0.99697304
       objectReference: {fileID: 0}
     - target: {fileID: 1068597257678563727, guid: ef1dd0845aa6a4cef9e999043447d039, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -12.489834
+      value: 22.10189
       objectReference: {fileID: 0}
     - target: {fileID: 1068597257678563727, guid: ef1dd0845aa6a4cef9e999043447d039, type: 3}
       propertyPath: m_LocalRotation.w
@@ -73818,15 +73818,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 2104673137817031364, guid: 0a57d2454786944f298dfc1ba1787e26, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -21.973328
+      value: 115.26825
       objectReference: {fileID: 0}
     - target: {fileID: 2104673137817031364, guid: 0a57d2454786944f298dfc1ba1787e26, type: 3}
       propertyPath: m_LocalPosition.y
-      value: -18.25877
+      value: 0.93237114
       objectReference: {fileID: 0}
     - target: {fileID: 2104673137817031364, guid: 0a57d2454786944f298dfc1ba1787e26, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -4.4130325
+      value: 30.178692
       objectReference: {fileID: 0}
     - target: {fileID: 2104673137817031364, guid: 0a57d2454786944f298dfc1ba1787e26, type: 3}
       propertyPath: m_LocalRotation.w
@@ -73862,7 +73862,7 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3376117400652952454, guid: 0a57d2454786944f298dfc1ba1787e26, type: 3}
       propertyPath: m_IsActive
-      value: 0
+      value: 1
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 0a57d2454786944f298dfc1ba1787e26, type: 3}
@@ -74605,7 +74605,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1690280945}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 137.24158, y: 19.191141, z: 34.591724}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children:
@@ -79344,17 +79344,13 @@ PrefabInstance:
   m_Modification:
     m_TransformParent: {fileID: 1690280946}
     m_Modifications:
-    - target: {fileID: 1126190618, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
-      propertyPath: data
-      value: 
-      objectReference: {fileID: 11400000, guid: 6c591b4d7375d4579807665bb2bcc69f, type: 2}
-    - target: {fileID: 1126190619, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
-      propertyPath: data
-      value: 
-      objectReference: {fileID: 11400000, guid: b19edce2b991d45c7a6cfaec097ec8b9, type: 2}
     - target: {fileID: 412854963691917426, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
       propertyPath: m_Name
       value: GuildguardSkeleton
+      objectReference: {fileID: 0}
+    - target: {fileID: 412854963691917426, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
+      propertyPath: m_IsActive
+      value: 0
       objectReference: {fileID: 0}
     - target: {fileID: 3915295574766006576, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
       propertyPath: m_RootOrder
@@ -79362,15 +79358,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 3915295574766006576, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -15.04
+      value: 122.20158
       objectReference: {fileID: 0}
     - target: {fileID: 3915295574766006576, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
       propertyPath: m_LocalPosition.y
-      value: -18.194168
+      value: 0.99697304
       objectReference: {fileID: 0}
     - target: {fileID: 3915295574766006576, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -11.8
+      value: 22.791725
       objectReference: {fileID: 0}
     - target: {fileID: 3915295574766006576, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
       propertyPath: m_LocalRotation.w
@@ -79400,10 +79396,6 @@ PrefabInstance:
       propertyPath: m_LocalEulerAnglesHint.z
       value: 0
       objectReference: {fileID: 0}
-    - target: {fileID: 5965740929030778247, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
-      propertyPath: behaviourTree
-      value: 
-      objectReference: {fileID: 11400000, guid: be73565d44d1f4e7cab99e8b770245b6, type: 2}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 05ba178005d7a4f12bcb9474593bb5c1, type: 3}
 --- !u!1001 &1805209596
@@ -82536,15 +82528,15 @@ PrefabInstance:
       objectReference: {fileID: 0}
     - target: {fileID: 5660498856279433209, guid: 722e47c41d4fa485080bce801bd9c64c, type: 3}
       propertyPath: m_LocalPosition.x
-      value: -9.17157
+      value: 128.07
       objectReference: {fileID: 0}
     - target: {fileID: 5660498856279433209, guid: 722e47c41d4fa485080bce801bd9c64c, type: 3}
       propertyPath: m_LocalPosition.y
-      value: -18.175047
+      value: 1.0160942
       objectReference: {fileID: 0}
     - target: {fileID: 5660498856279433209, guid: 722e47c41d4fa485080bce801bd9c64c, type: 3}
       propertyPath: m_LocalPosition.z
-      value: -1.1917229
+      value: 33.4
       objectReference: {fileID: 0}
     - target: {fileID: 5660498856279433209, guid: 722e47c41d4fa485080bce801bd9c64c, type: 3}
       propertyPath: m_LocalRotation.w
